+++ YouTubeBrowser/Story/WebViewController.swift
@@ -10,29 +10,70 @@ import Cocoa
 import WebKit
 
 class WebViewController: NSViewController {
-  var mainPresenter: MainPresenterProtocol!
+  // MARK: Dependencies
+
+  var mainPresenter: MainPresenterProtocol?
+
+  // MARK: @IBOutlets
 
   @IBOutlet private var webView: WKWebView!
+
+  // MARK: Life cycle
 
   override func viewDidLoad() {
     super.viewDidLoad()
 
     setupWebView()
-
-    mainPresenter = SwinjectAssembler.resolve(MainPresenterProtocol.self)
-    mainPresenter.webView = self
   }
 
   private func setupWebView() {
     webView.navigationDelegate = self
-    webView.configuration.mediaTypesRequiringUserActionForPlayback = .all
-    webView.customUserAgent = Constants.WebKit.userAgent
+
+    webView.customUserAgent =
+      Constants.WebKit.userAgent
+
+    webView.configuration.mediaTypesRequiringUserActionForPlayback =
+      Constants.WebKit.mediaTypesRequiringUserActionForPlayback
   }
 }
 
+// MARK: - WKNavigationDelegate
+
 extension WebViewController: WKNavigationDelegate {
-  func webView(_ webView: WKWebView, didFinish navigation: WKNavigation!) {}
+  func webView(_ webView: WKWebView, didFinish navigation: WKNavigation!) {
+    // inject javascript only once after page load
+    injectJavascript()
+  }
+
+  private func injectJavascript() {
+    // js functions to cleanup YouTube toolbar from clutter stuff not related to out app
+    let cleanUiJsUrl = Bundle.main.url(forResource: "CleanUiFunctions", withExtension: "js")!
+    let cleanUiJs = try! String(contentsOf: cleanUiJsUrl)
+
+    // js functions to handle player stuff (mini player, fullscreen, autoplay, etc.)
+    let playerJsUrl = Bundle.main.url(forResource: "PlayerFunctions", withExtension: "js")!
+    let playerJs = try! String(contentsOf: playerJsUrl)
+
+    // inject file used to execute after page load
+    let injectJsUrl = Bundle.main.url(forResource: "Inject", withExtension: "js")!
+    let injectJs = try! String(contentsOf: injectJsUrl)
+
+    // join into single js file
+    let js = [cleanUiJs, playerJs, injectJs].joined(separator: "\n")
+
+    // execute and hope
+    webView.evaluateJavaScript(js) { _, error in
+      if let error = error {
+        // TODO: Show execution error!
+        print(error)
+      } else {
+        print("Javascript injected!")
+      }
+    }
+  }
 }
+
+// MARK: - WebViewProtocol
 
 extension WebViewController: WebViewProtocol {
   func navigateToUrl(url: URL) {